requirements.txt
Flask==2.1.1
tensorflow==2.8.0
Pillow==9.0.1
protobuf==3.20.0